Refactor styles organization; move styles files to app feature directory and update import paths for improved structure and maintainability.
