+++ lib/ui/features/companies/widgets/companies_screen.dart
@@ -4,8 +4,8 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
 
 import '../../../../config/service_locator/service_locator_provider.dart';
-import '../../../styles/styles.dart';
 import '../../app/routing/routes.dart';
+import '../../app/styles/styles.dart';
 import '../companies_view_model.dart';
 import 'company_tile.dart';
